Add HLIR (High Level Intermediate Representation)
Created a representation of a built package that is more ergonomic to
work with in future stages of compilation.

The HLIR extracts the present function prototypes (external and defined)
before structuring the code into a set of functions which are then
traversed and has the variables turned into integer identifiers which
allows for the redefinition of variables.

Signed-off-by: Jacob Evans <jacob@evansj.dev>